refactor(edge/upgrade): move form to component
