+++ app/react/portainer/environments/update-schedules/CreateView/CreateView.tsx
@@ -1,5 +1,5 @@
 import { Settings } from 'react-feather';
-import { Field, Formik, Form as FormikForm, useFormikContext } from 'formik';
+import { Formik } from 'formik';
 import { useRouter } from '@uirouter/react';
 
 import { notifySuccess } from '@/portainer/services/notifications';
@@ -8,19 +8,15 @@ import {
   FeatureFlag,
 } from '@/portainer/feature-flags/useRedirectFeatureFlag';
 
-import { FormControl } from '@@/form-components/FormControl';
-import { Input } from '@@/form-components/Input';
 import { PageHeader } from '@@/PageHeader';
 import { Widget } from '@@/Widget';
-import { LoadingButton } from '@@/buttons';
-import { NavTabs } from '@@/NavTabs';
 
 import { ScheduleType } from '../types';
 import { useCreateMutation } from '../queries/create';
 
 import { FormValues } from './types';
 import { validation } from './validation';
-import { EdgeGroupsField } from './EdgeGroupsField';
+import { Form } from './Form';
 
 const initialValues: FormValues = {
   name: '',
@@ -63,7 +59,7 @@ export function CreateView() {
                 validateOnMount
                 validationSchema={validation}
               >
-                <Form />
+                <Form isLoading={createMutation.isLoading} />
               </Formik>
             </Widget.Body>
           </Widget>
@@ -71,65 +67,4 @@ export function CreateView() {
       </div>
     </>
   );
-}
-
-function Form() {
-  const { setFieldValue, errors, values, isSubmitting, isValid } =
-    useFormikContext<typeof initialValues>();
-
-  return (
-    <FormikForm className="form-horizontal">
-      <FormControl
-        label="Name"
-        required
-        inputId="name-input"
-        errors={errors.name}
-      >
-        <Field as={Input} name="name" id="name-input" />
-      </FormControl>
-
-      <EdgeGroupsField />
-
-      <div className="form-group">
-        <div className="col-sm-12">
-          <NavTabs
-            options={[
-              {
-                id: ScheduleType.Upgrade,
-                label: 'Upgrade',
-                children: <UpgradeForm />,
-              },
-              {
-                id: ScheduleType.Rollback,
-                label: 'Rollback',
-                children: <RollbackForm />,
-              },
-            ]}
-            selectedId={values.type}
-            onSelect={(value) => setFieldValue('type', value)}
-          />
-        </div>
-      </div>
-
-      <div className="form-group">
-        <div className="col-sm-12">
-          <LoadingButton
-            disabled={!isValid}
-            isLoading={isSubmitting}
-            loadingText="Creating..."
-          >
-            Create Schedule
-          </LoadingButton>
-        </div>
-      </div>
-    </FormikForm>
-  );
-}
-
-function UpgradeForm() {
-  return <div />;
-}
-
-function RollbackForm() {
-  return <div />;
 }
